+++ src/routes/user/services/index.js
@@ -1,6 +1,9 @@
+const crypto = require('crypto')
 const Helpers = require('../../../helpers')
 const UserModel = require('../models/index')
+const mailer = require('../../../utils/mail')
 const { sign } = require('../../../utils/jwt')
+const { date } = require('joi')
 
 const UserServices = (() => {
   const register = async userData => {
@@ -17,7 +20,10 @@ const UserServices = (() => {
       })
     } catch (error) {
       return Helpers.triggerLoggerAndReturnResult(
-        { status: false, error },
+        {
+          status: false,
+          error: error.message
+        },
         'error'
       )
     }
@@ -31,6 +37,9 @@ const UserServices = (() => {
         .split(':')
 
       const [user] = await UserModel.find({ email, password })
+
+      if (!user) throw { message: 'User or email invalid.' }
+
       const token = sign({
         user: user.id
       })
@@ -43,7 +52,10 @@ const UserServices = (() => {
       })
     } catch (error) {
       return Helpers.triggerLoggerAndReturnResult(
-        { status: false, error },
+        {
+          status: false,
+          error: error.message
+        },
         'error'
       )
     }
@@ -57,10 +69,87 @@ const UserServices = (() => {
     })
   }
 
+  const forgotPassword = async userEmail => {
+    try {
+      const [, hash] = userEmail.split(' ')
+      const [email] = Buffer.from(hash, 'base64').toString().split(':')
+      const user = await UserModel.findOne({ email })
+
+      if (!user) throw { message: 'Email not registered' }
+
+      const token = crypto.randomBytes(32).toString('hex')
+      const now = new Date()
+      now.setHours(now.getHours() + 1)
+
+      await UserModel.findByIdAndUpdate(user.id, {
+        '$set': {
+          passwordResetToken: token,
+          passwordResetExpires: now
+        }
+      })
+
+      await mailer.sendMail({
+        name: user.userName,
+        link: `/forgot-password/${user.id}/${token}`,
+        test: true,
+      })
+
+      return Helpers.triggerLoggerAndReturnResult({
+        status: true,
+        data: user,
+        message: 'A email will be send to you to reset your email!'
+      })
+    } catch (error) {
+      return Helpers.triggerLoggerAndReturnResult(
+        {
+          status: false,
+          error
+        },
+        'error'
+      )
+    }
+  }
+
+  const resetPassword = async userData => {
+    try {
+      const { userId, token, password } = userData
+      const now = new Date()
+      const user = await UserModel.findById(userId)
+        .select('passwordResetToken passwordResetExpires')
+
+      if (!user) throw { message: 'Email not registered' }
+
+      if (token !== user.passwordResetToken) throw { message: 'Token invalid' }
+
+      if (now > user.passwordResetExpires) throw { 
+        message: 'Token expired, generate new one' }
+
+      user.password = password
+
+      await user.save()
+
+      return Helpers.triggerLoggerAndReturnResult({
+        status: true,
+        message: 'New password saved with success!'
+      })
+
+    } catch (error) {
+      return Helpers.triggerLoggerAndReturnResult(
+        {
+          status: false,
+          error
+        },
+        'error'
+      )
+    }
+  }
+
   return {
     register,
     signin,
-    me
+    me,
+    forgotPassword,
+    resetPassword
   }
 })()