+++ src/utils/logger/index.js
@@ -1,8 +1,0 @@
-const pino = require('pino')({
-  prettyPrint: {
-    levelFirst: true,
-    colorize: true
-  }
-})
-
-module.exports = pino