+++ src/utils/mail/index.js
@@ -1,0 +1,22 @@
+const sendgrid = require('./mailSenders/sendgrid')
+const mailtrap = require('./mailSenders/mailtrap')
+
+const initialConfig = {
+  test: false
+}
+
+const mailer = (() => {
+  const sendMail = async (config) => { 
+    if (!config.test) {
+      await sendgrid.send()
+      return
+    }
+    await mailtrap.send(config)
+  }
+
+  return {
+    sendMail,
+  }
+})()
+
+module.exports = mailer