+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), even though this is not a library, so there will eventually be breaking changes in minor version updates, managed by the team.
 
+## [0.3.0] - 2021-08-12
+### Added
+- Create a router and service to reset password
+- Add a mailer to mailtrap and sendgrid
+- Add lib to validate data called JOI.js
+- Add view engine handlebars for emails
+- Add lib Nodemailer to send email to users
 ## [0.2.0] - 2021-08-12
 ### Added
 - Create revenue router, controller, service and schema